Prevent deserialization of the original request when validating request id

+++ rpc_sidecar/src/lib.rs
@@ -118,7 +118,7 @@ fn resolve_address(address: &str) -> anyhow::Result<SocketAddr> {
         .ok_or_else(|| anyhow::anyhow!("failed to resolve address"))
 }
 
-fn encode_request(req: &BinaryRequest, id: u64) -> Result<Vec<u8>, bytesrepr::Error> {
+fn encode_request(req: &BinaryRequest, id: u16) -> Result<Vec<u8>, bytesrepr::Error> {
     let header = BinaryRequestHeader::new(SUPPORTED_PROTOCOL_VERSION, req.tag(), id);
     let mut bytes = Vec::with_capacity(header.serialized_length() + req.serialized_length());
     header.write_bytes(&mut bytes)?;
